requirements.txt
streamlit==1.38.0
google-generativeai==0.7.2
pypdf==4.3.1
PyPDF2==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
